Add files via upload
Responsive E - Commerce Website in Django with MySQL

SHOPPERSSITE Responsive E-Commerce Website is Developed using HTML & CSS, Javascript, Bootstrap. It provides the user with a list of the various products available for purchase in the store. For the convenience of online shopping, a shopping cart is provided to the user. The system is implemented using Python’s web framework Django.